+++ CHANGELOG.md
@@ -5,6 +5,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Fixed
+- Fix path dependencies using project's workspace dependencies by @daimond113
+
 ## [0.6.0] - 2025-02-22
 ### Added
 - Improve installation experience by @lukadev-0